add HtmlClassNameProvider component + make it work for docs versions
